Refactored gcc-compiler-generator and created a clang one
